Document env setup for Firebase
