+++ .env.example
@@ -1,0 +1,10 @@
+# Example environment variables for Firebase
+WEB_API_KEY=
+ANDROID_API_KEY=
+IOS_API_KEY=
+MACOS_API_KEY=
+WINDOWS_API_KEY=
+PROJECT_ID=
+APP_ID=
+MESSAGING_SENDER_ID=
+STORAGE_BUCKET=
