+++ .gitignore
@@ -43,3 +43,8 @@ app.*.map.json
 /android/app/debug
 /android/app/profile
 /android/app/release
+
+# Local Firebase configuration
+lib/firebase_options.dart
+android/app/google-services.json
+.env
