+++ android/app/google-services.json.example
@@ -1,21 +1,21 @@
 {
   "project_info": {
-    "project_number": "235506639228",
-    "project_id": "oouchistock",
-    "storage_bucket": "oouchistock.firebasestorage.app"
+    "project_number": "YOUR_PROJECT_NUMBER",
+    "project_id": "YOUR_PROJECT_ID",
+    "storage_bucket": "YOUR_STORAGE_BUCKET"
   },
   "client": [
     {
       "client_info": {
-        "mobilesdk_app_id": "1:235506639228:android:9b0d9327c4c52f8bdcc4f9",
+        "mobilesdk_app_id": "YOUR_MOBILESDK_APP_ID",
         "android_client_info": {
-          "package_name": "com.example.oouchi_stock"
+          "package_name": "YOUR_PACKAGE_NAME"
         }
       },
       "oauth_client": [],
       "api_key": [
         {
-          "current_key": "AIzaSyAZBXjGKa-kalaz9ysGGXpocyZf5qXgJFc"
+          "current_key": "YOUR_API_KEY"
         }
       ],
       "services": {
