--- lib/firebase_options.dart
@@ -1,88 +1,0 @@
-// File generated by FlutterFire CLI.
-// ignore_for_file: type=lint
-import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
-import 'package:flutter/foundation.dart'
-    show defaultTargetPlatform, kIsWeb, TargetPlatform;
-
-/// Default [FirebaseOptions] for use with your Firebase apps.
-///
-/// Example:
-/// ```dart
-/// import 'firebase_options.dart';
-/// // ...
-/// await Firebase.initializeApp(
-///   options: DefaultFirebaseOptions.currentPlatform,
-/// );
-/// ```
-class DefaultFirebaseOptions {
-  static FirebaseOptions get currentPlatform {
-    if (kIsWeb) {
-      return web;
-    }
-    switch (defaultTargetPlatform) {
-      case TargetPlatform.android:
-        return android;
-      case TargetPlatform.iOS:
-        return ios;
-      case TargetPlatform.macOS:
-        return macos;
-      case TargetPlatform.windows:
-        return windows;
-      case TargetPlatform.linux:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions have not been configured for linux - '
-          'you can reconfigure this by running the FlutterFire CLI again.',
-        );
-      default:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions are not supported for this platform.',
-        );
-    }
-  }
-
-  static const FirebaseOptions web = FirebaseOptions(
-    apiKey: 'AIzaSyC86h_aMuttbAujWWvfCycxE7M1Tbm4UzQ',
-    appId: '1:235506639228:web:d5930cfb2063f33bdcc4f9',
-    messagingSenderId: '235506639228',
-    projectId: 'oouchistock',
-    authDomain: 'oouchistock.firebaseapp.com',
-    storageBucket: 'oouchistock.firebasestorage.app',
-    measurementId: 'G-5TQ4VR5ST4',
-  );
-
-  static const FirebaseOptions android = FirebaseOptions(
-    apiKey: 'AIzaSyAZBXjGKa-kalaz9ysGGXpocyZf5qXgJFc',
-    appId: '1:235506639228:android:9b0d9327c4c52f8bdcc4f9',
-    messagingSenderId: '235506639228',
-    projectId: 'oouchistock',
-    storageBucket: 'oouchistock.firebasestorage.app',
-  );
-
-  static const FirebaseOptions ios = FirebaseOptions(
-    apiKey: 'AIzaSyDiEM98ZpsHfsoTSNperb2Qb-R01UZZCtY',
-    appId: '1:235506639228:ios:31f95ce3ae1edf25dcc4f9',
-    messagingSenderId: '235506639228',
-    projectId: 'oouchistock',
-    storageBucket: 'oouchistock.firebasestorage.app',
-    iosBundleId: 'com.example.oouchiStock',
-  );
-
-  static const FirebaseOptions macos = FirebaseOptions(
-    apiKey: 'AIzaSyDiEM98ZpsHfsoTSNperb2Qb-R01UZZCtY',
-    appId: '1:235506639228:ios:31f95ce3ae1edf25dcc4f9',
-    messagingSenderId: '235506639228',
-    projectId: 'oouchistock',
-    storageBucket: 'oouchistock.firebasestorage.app',
-    iosBundleId: 'com.example.oouchiStock',
-  );
-
-  static const FirebaseOptions windows = FirebaseOptions(
-    apiKey: 'AIzaSyC86h_aMuttbAujWWvfCycxE7M1Tbm4UzQ',
-    appId: '1:235506639228:web:b4687a360896cd93dcc4f9',
-    messagingSenderId: '235506639228',
-    projectId: 'oouchistock',
-    authDomain: 'oouchistock.firebaseapp.com',
-    storageBucket: 'oouchistock.firebasestorage.app',
-    measurementId: 'G-R16JC1VSVH',
-  );
-}
